добавлен паттерн Наблюдатель и выделены функции для рандома
