+++ src/utils/common.js
@@ -1,13 +1,3 @@
-//получить случайный элемент массива
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
-
-//получить случайное целочисленное число
-function getRandomInteger(min, max) {
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
-
 //получить названия пунктов назначения
 const getDestinationNames = (destinations, points = []) => {
   if (points.length > 0) {
@@ -40,21 +30,6 @@ function getElementById(elements, itemsId) {
   return elements.find((element) => element.id === itemsId);
 }
 
-//перемешать массив
-function shuffle(items) {
-  for (let i = items.length - 1; i > 0; i--) {
-    const randomIndex = Math.floor(Math.random() * (i + 1));
-    [items[i], items[randomIndex]] = [items[randomIndex], items[i]];
-  }
-  return items;
-}
-
-//изменить рандомно размер массива
-function changeLengthRandom(items) {
-  const max = getRandomInteger(0, items.length - 1);
-  return items.slice(0, max);
-}
-
 //обновление элементов в массиве объектов
 function updateItem(items, update) {
   return items.map((item) => item.id === update.id ? update : item);
@@ -64,15 +39,11 @@ function updateItem(items, update) {
 const sortByPrice = (a, b) => b.basePrice - a.basePrice;
 
 export {
-  getRandomArrayElement,
-  getRandomInteger,
   getDestinationNames,
   capitalize,
   getFullPrice,
   getElementByType,
   getElementById,
-  shuffle,
-  changeLengthRandom,
   isEscape,
   updateItem,
   sortByPrice,
